tests/requirements-test.txt
pytest
httpx
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
pydantic
pydantic-settings
fastapi
pillow
piexif
python-dotenv
